+++ CHANGES.md
@@ -7,3 +7,5 @@ Change Log
 * changed asCanvas(ratio: Double = 1) to take a "window" ratio value, default=1.
 * changed all strokeWidth to be Double
 * added Outline 
+* added addIconParts(function: js.Function)
+* added addLetterSIDCicons(function: js.Function), addNumberSIDCicons(function: js.Function)